+++ projects/batfish/src/test/resources/org/batfish/grammar/palo_alto/testconfigs/drop-inter-vsys-missing-external-egress
@@ -10,6 +10,7 @@ set vsys vsys1 zone vsys1-vsys2 network external vsys2
 set vsys vsys2 import network interface ethernet1/2
 set vsys vsys2 import visible-vsys vsys1
 set vsys vsys2 zone zout network layer3 ethernet1/2
+! missing external zone pointing to vsys1
 set vsys vsys1 rulebase security rules rule1 from any
 set vsys vsys1 rulebase security rules rule1 to any
 set vsys vsys1 rulebase security rules rule1 action allow
