+++ projects/batfish/src/test/resources/org/batfish/grammar/palo_alto/testconfigs/drop-intra-vsys-next-vr-missing-egress
@@ -5,6 +5,7 @@ set network virtual-router vr1 interface ethernet1/1
 set network virtual-router vr1 routing-table ip static-route sr1 destination 0.0.0.0/0
 set network virtual-router vr1 routing-table ip static-route sr1 nexthop next-vr vr2
 set network virtual-router vr2 interface ethernet1/2
+! return traffic not routable because missing corresponding next-vr route back to vr1
 set vsys vsys1 import network interface ethernet1/1
 set vsys vsys1 import network interface ethernet1/2
 set vsys vsys1 zone z network layer3 ethernet1/1
