+++ projects/batfish/src/test/resources/org/batfish/grammar/palo_alto/testconfigs/drop-intra-vsys-next-vr-missing-ingress
@@ -2,6 +2,7 @@ set deviceconfig system hostname drop-intra-vsys-next-vr-missing-ingress
 set network interface ethernet ethernet1/1 layer3 ip 10.0.1.1/24
 set network interface ethernet ethernet1/2 layer3 ip 10.0.2.1/24
 set network virtual-router vr1 interface ethernet1/1
+! forward traffic not routable because missing corresponding next-vr route to vr2
 set network virtual-router vr2 interface ethernet1/2
 set network virtual-router vr2 routing-table ip static-route sr2 destination 0.0.0.0/0
 set network virtual-router vr2 routing-table ip static-route sr2 nexthop next-vr vr1
